[WIP] display: use integer position
maybe it's not needed. it adds an overhead

+++ src/drivers/Display.cpp
@@ -19,22 +19,22 @@ void Display::send()
 	hwDisplaySend();
 }
 
-void Display::draw(uint8_t x, uint8_t y, DisplayColor color)
+void Display::draw(int x, int y, DisplayColor color)
 {
 	hwDisplayDraw(x, y, color);
 }
 
-void Display::drawImage(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t* image)
+void Display::drawImage(int x, int y, uint8_t w, uint8_t h, const uint8_t* image)
 {
 	hwDisplayDrawImage(x, y, w, h, image);
 }
 
-void Display::drawImage(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint16_t* image)
+void Display::drawImage(int x, int y, uint8_t w, uint8_t h, const uint16_t* image)
 {
 	hwDisplayDrawImage(x, y, w, h, image);
 }
 
-void Display::drawRectangle(uint8_t x, uint8_t y, uint8_t width, uint8_t height, DisplayColor color)
+void Display::drawRectangle(int x, int y, uint8_t width, uint8_t height, DisplayColor color)
 {
 	hwDisplayDrawRectangle(x, y, width, height, color);
 }
