+++ src/drivers/Display.h
@@ -13,10 +13,10 @@ public:
 
   void send();
 
-  void draw(uint8_t x, uint8_t y, DisplayColor color);
-  void drawImage(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t* image);
-  void drawImage(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint16_t* image);
-  void drawRectangle(uint8_t x, uint8_t y, uint8_t width, uint8_t height, DisplayColor color);
+  void draw(int x, int y, DisplayColor color);
+  void drawImage(int x, int y, uint8_t w, uint8_t h, const uint8_t* image);
+  void drawImage(int x, int y, uint8_t w, uint8_t h, const uint16_t* image);
+  void drawRectangle(int x, int y, uint8_t width, uint8_t height, DisplayColor color);
   void fill(DisplayColor color);
 
   uint8_t width();
